+++ environ/environ.py
@@ -154,15 +154,24 @@ class Env:
     _EMAIL_BASE_OPTIONS = ['EMAIL_USE_TLS', 'EMAIL_USE_SSL']
 
     DEFAULT_SEARCH_ENV = 'SEARCH_URL'
+
+    ELASTICSEARCH_FAMILY = [
+        'elasticsearch',
+        'elasticsearch2',
+        'elasticsearch5',
+    ]
+
     SEARCH_SCHEMES = {
-        "elasticsearch": "haystack.backends.elasticsearch_backend."
-                         "ElasticsearchSearchEngine",
-        "elasticsearch2": "haystack.backends.elasticsearch2_backend."
-                          "Elasticsearch2SearchEngine",
-        "solr": "haystack.backends.solr_backend.SolrEngine",
-        "whoosh": "haystack.backends.whoosh_backend.WhooshEngine",
-        "xapian": "haystack.backends.xapian_backend.XapianEngine",
-        "simple": "haystack.backends.simple_backend.SimpleEngine",
+        'elasticsearch': 'haystack.backends.elasticsearch_backend.'
+                         'ElasticsearchSearchEngine',
+        'elasticsearch2': 'haystack.backends.elasticsearch2_backend.'
+                          'Elasticsearch2SearchEngine',
+        'elasticsearch5': 'haystack.backends.elasticsearch5_backend.'
+                          'Elasticsearch5SearchEngine',
+        'solr': 'haystack.backends.solr_backend.SolrEngine',
+        'whoosh': 'haystack.backends.whoosh_backend.WhooshEngine',
+        'xapian': 'haystack.backends.xapian_backend.XapianEngine',
+        'simple': 'haystack.backends.simple_backend.SimpleEngine',
     }
 
     def __init__(self, interpolate=False, **scheme):
@@ -654,7 +663,7 @@ class Env:
         # check commons params
         params = {}
         if url.query:
-            params = parse_qs(url.query)
+            params = parse_qs(url.query)  # type: dict
             if 'EXCLUDED_INDEXES' in params.keys():
                 config['EXCLUDED_INDEXES'] = \
                     params['EXCLUDED_INDEXES'][0].split(',')
@@ -672,7 +681,7 @@ class Env:
         if url.scheme == 'simple':
             return config
 
-        if url.scheme in ['solr', 'elasticsearch', 'elasticsearch2']:
+        if url.scheme in ['solr'] + cls.ELASTICSEARCH_FAMILY:
             if 'KWARGS' in params.keys():
                 config['KWARGS'] = params['KWARGS'][0]
 
@@ -688,8 +697,7 @@ class Env:
                 config['TIMEOUT'] = cls.parse_value(params['TIMEOUT'][0], int)
             return config
 
-        if url.scheme in ['elasticsearch', 'elasticsearch2']:
-
+        if url.scheme in cls.ELASTICSEARCH_FAMILY:
             split = path.rsplit("/", 1)
 
             if len(split) > 1: